bump shell pkg + remove import of v-clean directive + use v-clean-html instead of v-html + working on install elemental operator from UI
