+++ pkg/elemental/l10n/en-us.yaml
@@ -1,7 +1,7 @@
 product:
   elemental: OS Management
   description: "Elemental is a software stack enabling a centralized, full cloud-native OS management with Kubernetes. <br><br> Cluster Node OSes are built and maintained via container images through the Elemental Toolkit and installed on new hosts using the Elemental CLI. <br><br> The Elemental Operator and the Rancher System Agent enable Rancher Manager to fully control Elemental clusters, from the installation and management of the OS on the Nodes to the provisioning of new K3s or RKE2 clusters in a centralized way."
-  notInstalledOrNoSchema: Either the user doesn't have enough permissions to run the OS Management extension or the Elemental Operator is not installed (required to run the OS Management extension). <br><br> For user permissions, check with your Rancher administrator if the correct role is assigned. To install Elemental Operator please follow the instructions on the official <a target="blank" href="https://elemental.docs.rancher.com/quickstart-cli#install-elemental-operator" target="_blank" rel="noopener nofollow">documentation</a>.
+  notInstalledOrNoSchema: Either the user doesn't have enough permissions to run the OS Management extension or the Elemental Operator is not installed (required to run the OS Management extension). <br><br> For user permissions, check with your Rancher administrator if the correct role is assigned. To install Elemental Operator please click the button below.
   notInstalledHasSchema: The Elemental Operator is required to run the OS Management extension. To install it please follow the instructions on the official <a target="blank" href="https://elemental.docs.rancher.com/quickstart-cli#install-elemental-operator" target="_blank" rel="noopener nofollow">documentation</a>.
 
 cluster:
@@ -69,6 +69,10 @@ typeLabel:
   elementalClusters: Clusters
 
 asyncButton:
+  elementalRepository:
+    action: Add Elemental Repository
+    success: Added
+    waiting: Adding&hellip;
   downloadMachineReg:
     action: Download Configuration File
     waiting: Preparing..
@@ -87,6 +91,12 @@ description:
   elementalClusters: Clusters managed by Elemental
 
 elemental:
+  appInstall:
+    button: Install Elemental Operator
+    reload: Unable to fetch Elemental Operator Helm chart - reload required
+  versionError:
+      title: Chart Version not found.
+      message: Unable to determine the latest stable version of the elemental chart. Please make sure the Helm repository
   menuLabels:
     dashboard: Dashboard
     titleDashboard: OS Management Dashboard
